requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
SQLAlchemy==2.0.31
Werkzeug==3.0.3
APScheduler==3.10.4
python-dotenv==1.0.1
pytest==8.3.2
email-validator==2.1.1
